Only mark the HEAD of PR pending if has fixup commits
This fixes an issue where if a new commit is added on top of fixup
commits that are marked as pending, github still thinks that things are
fine, as the last commits statuses are the ones that are taken into
account for the overall PR status.

We don't have to worry about cleaning up these pending statuses, because
they will always be either on or after fixup and squash commits. This
means that after `!squash` they will get new SHAs and therefore the
statuses will be forgotten.