Mark PR head's satus as peer reviewed on +1 comment
This is to be able to force PRs to be peer reviewed before merging.
This, however, is a really simple solution. It only marks the current
latest commit as reviewed and therefore comments like: "+1 once you fix
that typo" will not mark the later commits (the typo fix) as peer
reviewed. Also, running `!squash` will most likely remove the peer
review status as the latest commit will get a new SHA if anything gets
squashed.