Add 'merging' label to PR on !merge
This is so that when the repository is configured such that all PRs have
to be up to date to be merged then we could track pushes to master and
keep updating all the PRs that have been marked with 'merging'. Keeping
the PRs up to date this way ensures that the PR can be merged as soon as
all necessary checks on it's latest commit have been run.