Update !merge command to try and merge the PR
This leaves a lot of the failure cases uncovered, but is a good starting
point. Further improvements include getting an updated ref and retrying
in case the head branch has been modified between when it was fetched
and when the merge request was sent and trying to detect other
recoverable failure scenarios.