Squash on !merge if has pending squash status
This means that it won't currently actually get merged, but once we set
up listeners for status changes etc for PRs with the 'merging' status,
this will not be a problem. Because we'll get notified when the
'success' review/squash status is set after the PR synchronized event
and we can check if the PR is ready for merging and merge then.