Update get/set status methods only work with PRs
This now enforces only the actual PR's latest commit's SHA being used
instead of the branch ref. This is the better approach, because then we
always have a consistent view of the PR and the statuses, even if it's
outdated. E.g. if more commits are added to the PR between when we get
the PR and when we get the statuses, then we might miss information
about the new commits, but at least we will not be, for example,
changing statuses on the new commits which we any information about.