Quickfix rebasing onto old revisions
There's an issue where the local repository used by the bot and more
specifically it's local branches that are tracking remote (origin)
branches become out of date. I.e. when updating the repository, we
currently only do a fetch, which means that if origin/master moves, we
don't move the local master with it. Which in turn means that if we do a
rebase against master than we're rebasing against that outdated ref.

This is a quickfix around the issue, that works by only using the remote
refs. A proper solution has to correctly handle all remotes (forks etc).
Although, even with forks, the base remote will always still be origin.