Update for github library changes
The changes to pointers are straight forward. The addition of
PullRequestOptions means the bot has to decide whether or not it wants
to squash all commits in the PR into one. Seeing as that would defeat a
lot of this purpose, the answer is no - it will not squash these commits
together.