Add a dockerfile
The PORT environment variable is defined in the image, because the user
can already specify the host port onto which the container's port 80
maps with the -p flag of `docker run`, so the extra configurability
wouldn't add much.

The SSH known_hosts and id_rsa files are expected to be available at
/etc/secret-volume because that way this image can easily be used in a
Kubernetes environment where secrets (such as SSH keys) can be mounted
as a volume to a container.