Parse head branch info from PR event webhook payload
Currently we're making an extra GET request against the API to get this
information, although we already have it available. This is first,
wasteful, and second, subject to a race condition we've observed. So
although the PR is definitely available when this webhook is triggered,
the GitHub API can still return a 404 for when requesting data for the
PR.

The google go-github library actually provides functions for parsing all
webhook events, so we don't necessarily have to implement this
ourselves. This functionality has either been recently added or wasn't
noticed during the initial development. Now changing the code not use
this own implementation of parsing is not as straight forward. Could be
considered as a future possibility.