Add a !check command for manually forcing PR checks
As explained in the README, this can sometimes be very valuable. Also,
right now there's no way to force a recheck after the bot has been dead
for a while, this provides some way of bringing PRs that were
created/updated during that downtime up to date.

I recently observed one particular scenario where this will be
invaluable. We've recently added retry mechanisms for when an endpoint
returns 404 when requested too soon. This helps in scenarios where the
resource was recently created, but is helpless against the following
scenario, which is (as far as I can see) impossible to detect
automatically without adding forced delays. The scenario:

- A person creates a PR
- The person quickly realizes that they made a mistake in the (last or
  only) commit and just as quickly amends the commit
- The person force pushes their updated commit (to avoid public shaming
  from their peers for the horrible mistake they'd done)
- The bot gets 2 pull_request synchronize events
- For the latter event, it pulls the PR info from GitHub API which
  hasn't been updated yet
- The bot now continues evaluating the old commit, that isn't any longer
  a part of the PR and setting statuses for which is useless
- The person sees that the bot isn't setting the status for their new
  commit and complains either directly to whoever is managing the bot
  or, worse, publicly on the PR