Try merging PRs after status updates
With this change the `!merge` command should now be reliably usable.
Because the bot is already listening to PR 'synchronize' events and is
changing the status of the PR events, then we're also effectively
looking for merge options on the PR 'synchronize' events as well.

There is an option where the `!merge` command has been entered, but the
PR doesn't have any approved reviews and so can't be merged. To reliably
have the PR merged in these cases, we should also listen to the review
events, but because this is not too likely a scenario, the command can
already be used without listening to these events.

Another QoL improvement would be to remove the 'merging' label when a PR
with that label has its status set to 'failed'. And then notify the
author of the PR, that they should take action. This too will be done
in the future.