Separate squash conflict errors from others in the git package
This way other errors returned from the same function will not be
treated as squash conflicts, but rather as internal errors. This is good
now, but even better soon when the method will have to include more
operations.

Not really liking how the errors are handled right now. They're being
logged at different levels and sometimes wrapped and sometimes not. It's
never clear if an error has been handled already or not. Could consider
taking [Dave Cheney's error wrapping approach][1] by using the
[github.com/pkg/errors][2] library sometime in the future.

[1]: https://dave.cheney.net/paste/gocon-spring-2016.pdf
[2]: https://github.com/pkg/errors