Fix squash race condition
ForcePushHeadTo, as the name says, pushed the HEAD to the specified
destination ref. This means, that for the combination of
RebaseAutosquash + ForcePushHeadTo to work correctly, nothing must
change the HEAD after RebaseAutosquash was run. Currently there's a
chance that something else will acquire the lock between the 2
functions are run, which could create considerable confusion and harm if
it were to happen.

This solution holds the lock until both commands are finished.