Add a timeout for the HTTP client
We don't want to stay waiting indefinitely for the GitHub API when it's
down, which is the default behavior if the Timeout for http.Client is
not specified. 30 seconds should be high enough a value not to cut off
slow requests, but to also avoid system hanging when GitHub API is
unresponsive.