Delay GitHub issue searches
Experience has shown that these searches can return outdated results
if run immediately on status update webhook events. This has caused some
PRs to get stuck in the merging state without being merged, because the
last status update event has already been sent, but the search during
that event did not show that the PR was ready to be merged. Because
there will be no additional events for that PR (unless users manually
use `!check`, for example) the PR will not get merged.

A simpler solution would have been to simply sleep for the specified
amount (2 seconds by default) and then do the search. But that approach
would mean that the GitHub webhook HTTP request and its connection is
unnecessarily kept open for the duration of the sleep + how long it then
takes to finish the search + merge. This is not too much of an issue for
the default 2 second delay, because Go's standard HTTP server handles
concurrency nicely, and even thousands of concurrent requests (even if
they're just sleeping) could be handled without a problem. However, if
it turns out that 2 seconds isn't sufficient, and a longer period has to
be used, then the described problem can start becoming an issue.
Furthermore, the async approach is much nicer towards GitHub webhook
service, which probably has to worry about too many connections that
stay open for too long already.

This approach does make the HTTP responses and tests a bit less
valuable, because neither can capture the final success/failure any
longer. Tests can still check that the async code sends all the right
requests to GitHub etc, but without parsing the logs or mocking the
delay mechanism, there's no way for the tests to verify the result.