Start scheduled operations immediately on SIGINT
We're using the graceful package for running the HTTP server. The
package [describes][1] the last step of it's process with "Returns from
the function, allowing the server to terminate." After the HTTP server
returns, we're currently waiting for all scheduled operations to finish
and then return from the main method.

With this change, we don't block the process from dying for too long
after the interrupt signal by listening to it and then making our best
effort to finish gracefully and fast. Finishing gracefully (allowing the
scheduled operations to finish, instead of unscheduling/stopping them)
is important, because the bot is currently built so that it can only
guarantee smooth operation if it catches and handles all the events it's
subscribed to.

The [docs][2] have the following to say about calling Notify multiple
times, which we do here - we call it for every delayed operation:

> It is allowed to call Notify multiple times with different channels
> and the same signals: each channel receives copies of incoming signals
> independently.

[1]: https://github.com/tylerb/graceful#behaviour
[2]: https://golang.org/pkg/os/signal/#Notify