Add git command for deleting a remote branch
GitHub API does not support deleting the branch after it's been merged.
This means that unless people go back and manually delete the branches
after using `!merge`, they won't be deleted. Forcing the people to go
back and manually do anything on a successful merge, however, goes
against the whole idea of having a `!merge` command in the first place -
not having to wait for squashing and other status checks (e.g. CI) to
finish to merge a PR that's been reviewed.