Delete remote branch after merge
This keeps the repository clean from stale branches that have already
been merged into master anyway.

PRs that are across forks are ignored, because the bot is really acting
on the base repository, and so touching the fork repository in any way,
even if possible, seems wrong. Time will show if we need to do some
cleanup across forks as well.