Fix squash status check race condition
This fixes, for example, the scenario described in 6a37b24. Essentially,
when requesting commits for a PR then if the PR has been recently
updated then GitHub API may return an outdated list. This means that
checking for fixup commits in that list may return a false response.

One similar race condition was fixed in b421821, where the outdated list
of commits meant no commits, but it didn't fix the scenario where the
outdated list meant a whole different set of commits.

This is only implemented for PullRequestEvents and not for the `!check`
command (IssueCommentEvents), because PullRequestEvents contain the head
SHA that is guaranteed to be up to date, whereas the IssueCommentEvents
don't. But this should not be an issue, because the `!check` command is
sent by comments written by actual users - events which are decoupled
from the git/GitHub events which may change the commit list. The
decoupling means that the race condition is unlikely to happen in these
cases.