Add (empty) context to GitHub requests
The GitHub library has been updated to take context.Context as the first
argument for its every function that does an outgoing request. This is
good because it adds more control over the behavior of these requests,
but there's no real need for using it right now. To make the types
compatible, but to avoid actually implementing cancellations, etc, the
[context.TODO][1] function is used, which is provided for exactly that
purpose.

[1]: https://godoc.org/context#TODO