Fix head commit check relying on GitHub response order
GitHub returns the commits in chronological order, which means that the
assumption that the last commit is the HEAD commit was wrong. The
assumption works in most cases, but some rebasing can easily mess up the
ordering and break the current check.

The fix is to find the HEAD based on the topological ordering provided
by the commits' parent references.