Retry getting PR's commits asynchronously
It happens quite frequently, that all 4 synchronous tries to get a list
of commits for a PR get the same outdated view of commits (the
unexpected head check) causing the fixup commit check to fail, although
it would've succeeded just a few seconds later. Now the retries are
asynchronous, and if so configured, will use exponential-like backoff
making it less likely that checking for fixup commits fails due to a
GitHub API race condition.