Expand !merge command
The reson for this PR is that some applications for mobile phones
(example SwiftHawk for iOS) post an additional line when commenting and
the !merge command gets ignored.

Example https://github.com/salemove/ios-sdk/pull/157

To avoid this kind of behaviour, the logic of finding the merge comment
is changed to search for prefix insead of the whole comment.